+++ src/Language/AST/IntValueNode.php
@@ -9,6 +9,6 @@ class IntValueNode extends Node implements ValueNode
     /** @var string */
     public $kind = NodeKind::INT;
 
-    /** @var mixed */
+    /** @var string */
     public $value;
 }