+++ src/Language/Printer.php
@@ -56,6 +56,7 @@ use function json_encode;
 use function preg_replace;
 use function sprintf;
 use function str_replace;
+use function strlen;
 use function strpos;
 
 /**
@@ -93,6 +94,12 @@ class Printer
     {
     }
 
+    /**
+     * Traverse an AST bottom-up, converting all nodes to strings.
+     *
+     * That means the AST is manipulated in such a way that it no longer
+     * resembles the well-formed result of parsing.
+     */
     public function printAST($ast)
     {
         return Visitor::visit(
@@ -100,7 +107,7 @@ class Printer
             [
                 'leave' => [
                     NodeKind::NAME => static function (NameNode $node) : string {
-                        return '' . $node->value;
+                        return $node->value;
                     },
 
                     NodeKind::VARIABLE => static function (VariableNode $node) : string {
@@ -111,7 +118,7 @@ class Printer
                         return $this->join($node->definitions, "\n\n") . "\n";
                     },
 
-                    NodeKind::OPERATION_DEFINITION => function (OperationDefinitionNode $node) {
+                    NodeKind::OPERATION_DEFINITION => function (OperationDefinitionNode $node) : string {
                         $op           = $node->operation;
                         $name         = $node->name;
                         $varDefs      = $this->wrap('(', $this->join($node->variableDefinitions, ', '), ')');
@@ -120,7 +127,7 @@ class Printer
 
                         // Anonymous queries with no directives or variable definitions can use
                         // the query short form.
-                        return $name === null && ! $directives && ! $varDefs && $op === 'query'
+                        return $name === null && strlen($directives ?? '') === 0 && ! $varDefs && $op === 'query'
                             ? $selectionSet
                             : $this->join([$op, $this->join([$name, $varDefs]), $directives, $selectionSet], ' ');
                     },
@@ -137,7 +144,7 @@ class Printer
                         return $this->block($node->selections);
                     },
 
-                    NodeKind::FIELD => function (FieldNode $node) {
+                    NodeKind::FIELD => function (FieldNode $node) : string {
                         return $this->join(
                             [
                                 $this->wrap('', $node->alias, ': ') . $node->name . $this->wrap(
@@ -157,10 +164,12 @@ class Printer
                     },
 
                     NodeKind::FRAGMENT_SPREAD => function (FragmentSpreadNode $node) : string {
-                        return '...' . $node->name . $this->wrap(' ', $this->join($node->directives, ' '));
+                        return '...'
+                            . $node->name
+                            . $this->wrap(' ', $this->join($node->directives, ' '));
                     },
 
-                    NodeKind::INLINE_FRAGMENT => function (InlineFragmentNode $node) {
+                    NodeKind::INLINE_FRAGMENT => function (InlineFragmentNode $node) : string {
                         return $this->join(
                             [
                                 '...',
@@ -181,7 +190,7 @@ class Printer
                             . $node->selectionSet;
                     },
 
-                    NodeKind::INT => static function (IntValueNode $node) {
+                    NodeKind::INT => static function (IntValueNode $node) : string {
                         return $node->value;
                     },
 
@@ -189,7 +198,7 @@ class Printer
                         return $node->value;
                     },
 
-                    NodeKind::STRING => function (StringValueNode $node, $key) {
+                    NodeKind::STRING => function (StringValueNode $node, $key) : string {
                         if ($node->block) {
                             return $this->printBlockString($node->value, $key === 'description');
                         }
@@ -197,7 +206,7 @@ class Printer
                         return json_encode($node->value);
                     },
 
-                    NodeKind::BOOLEAN => static function (BooleanValueNode $node) {
+                    NodeKind::BOOLEAN => static function (BooleanValueNode $node) : string {
                         return $node->value ? 'true' : 'false';
                     },
 
@@ -226,6 +235,7 @@ class Printer
                     },
 
                     NodeKind::NAMED_TYPE => static function (NamedTypeNode $node) : string {
+                        // @phpstan-ignore-next-line the printer works bottom up, so this is already a string here
                         return $node->name;
                     },
 
@@ -237,7 +247,7 @@ class Printer
                         return $node->type . '!';
                     },
 
-                    NodeKind::SCHEMA_DEFINITION => function (SchemaDefinitionNode $def) {
+                    NodeKind::SCHEMA_DEFINITION => function (SchemaDefinitionNode $def) : string {
                         return $this->join(
                             [
                                 'schema',
@@ -252,11 +262,11 @@ class Printer
                         return $def->operation . ': ' . $def->type;
                     },
 
-                    NodeKind::SCALAR_TYPE_DEFINITION => $this->addDescription(function (ScalarTypeDefinitionNode $def) {
+                    NodeKind::SCALAR_TYPE_DEFINITION => $this->addDescription(function (ScalarTypeDefinitionNode $def) : string {
                         return $this->join(['scalar', $def->name, $this->join($def->directives, ' ')], ' ');
                     }),
 
-                    NodeKind::OBJECT_TYPE_DEFINITION => $this->addDescription(function (ObjectTypeDefinitionNode $def) {
+                    NodeKind::OBJECT_TYPE_DEFINITION => $this->addDescription(function (ObjectTypeDefinitionNode $def) : string {
                         return $this->join(
                             [
                                 'type',
@@ -269,7 +279,7 @@ class Printer
                         );
                     }),
 
-                    NodeKind::FIELD_DEFINITION => $this->addDescription(function (FieldDefinitionNode $def) {
+                    NodeKind::FIELD_DEFINITION => $this->addDescription(function (FieldDefinitionNode $def) : string {
                         $noIndent = Utils::every($def->arguments, static function (string $arg) : bool {
                             return strpos($arg, "\n") === false;
                         });
@@ -282,7 +292,7 @@ class Printer
                             . $this->wrap(' ', $this->join($def->directives, ' '));
                     }),
 
-                    NodeKind::INPUT_VALUE_DEFINITION => $this->addDescription(function (InputValueDefinitionNode $def) {
+                    NodeKind::INPUT_VALUE_DEFINITION => $this->addDescription(function (InputValueDefinitionNode $def) : string {
                         return $this->join(
                             [
                                 $def->name . ': ' . $def->type,
@@ -294,7 +304,7 @@ class Printer
                     }),
 
                     NodeKind::INTERFACE_TYPE_DEFINITION => $this->addDescription(
-                        function (InterfaceTypeDefinitionNode $def) {
+                        function (InterfaceTypeDefinitionNode $def) : string {
                             return $this->join(
                                 [
                                     'interface',
@@ -307,13 +317,13 @@ class Printer
                         }
                     ),
 
-                    NodeKind::UNION_TYPE_DEFINITION => $this->addDescription(function (UnionTypeDefinitionNode $def) {
+                    NodeKind::UNION_TYPE_DEFINITION => $this->addDescription(function (UnionTypeDefinitionNode $def) : string {
                         return $this->join(
                             [
                                 'union',
                                 $def->name,
                                 $this->join($def->directives, ' '),
-                                $def->types
+                                count($def->types ?? []) > 0
                                     ? '= ' . $this->join($def->types, ' | ')
                                     : '',
                             ],
@@ -321,7 +331,7 @@ class Printer
                         );
                     }),
 
-                    NodeKind::ENUM_TYPE_DEFINITION => $this->addDescription(function (EnumTypeDefinitionNode $def) {
+                    NodeKind::ENUM_TYPE_DEFINITION => $this->addDescription(function (EnumTypeDefinitionNode $def) : string {
                         return $this->join(
                             [
                                 'enum',
@@ -333,13 +343,13 @@ class Printer
                         );
                     }),
 
-                    NodeKind::ENUM_VALUE_DEFINITION => $this->addDescription(function (EnumValueDefinitionNode $def) {
+                    NodeKind::ENUM_VALUE_DEFINITION => $this->addDescription(function (EnumValueDefinitionNode $def) : string {
                         return $this->join([$def->name, $this->join($def->directives, ' ')], ' ');
                     }),
 
                     NodeKind::INPUT_OBJECT_TYPE_DEFINITION => $this->addDescription(function (
                         InputObjectTypeDefinitionNode $def
-                    ) {
+                    ) : string {
                         return $this->join(
                             [
                                 'input',
@@ -351,7 +361,7 @@ class Printer
                         );
                     }),
 
-                    NodeKind::SCHEMA_EXTENSION => function (SchemaTypeExtensionNode $def) {
+                    NodeKind::SCHEMA_EXTENSION => function (SchemaTypeExtensionNode $def) : string {
                         return $this->join(
                             [
                                 'extend schema',
@@ -362,7 +372,7 @@ class Printer
                         );
                     },
 
-                    NodeKind::SCALAR_TYPE_EXTENSION => function (ScalarTypeExtensionNode $def) {
+                    NodeKind::SCALAR_TYPE_EXTENSION => function (ScalarTypeExtensionNode $def) : string {
                         return $this->join(
                             [
                                 'extend scalar',
@@ -373,7 +383,7 @@ class Printer
                         );
                     },
 
-                    NodeKind::OBJECT_TYPE_EXTENSION => function (ObjectTypeExtensionNode $def) {
+                    NodeKind::OBJECT_TYPE_EXTENSION => function (ObjectTypeExtensionNode $def) : string {
                         return $this->join(
                             [
                                 'extend type',
@@ -386,7 +396,7 @@ class Printer
                         );
                     },
 
-                    NodeKind::INTERFACE_TYPE_EXTENSION => function (InterfaceTypeExtensionNode $def) {
+                    NodeKind::INTERFACE_TYPE_EXTENSION => function (InterfaceTypeExtensionNode $def) : string {
                         return $this->join(
                             [
                                 'extend interface',
@@ -398,13 +408,13 @@ class Printer
                         );
                     },
 
-                    NodeKind::UNION_TYPE_EXTENSION => function (UnionTypeExtensionNode $def) {
+                    NodeKind::UNION_TYPE_EXTENSION => function (UnionTypeExtensionNode $def) : string {
                         return $this->join(
                             [
                                 'extend union',
                                 $def->name,
                                 $this->join($def->directives, ' '),
-                                $def->types
+                                count($def->types ?? []) > 0
                                     ? '= ' . $this->join($def->types, ' | ')
                                     : '',
                             ],
@@ -412,7 +422,7 @@ class Printer
                         );
                     },
 
-                    NodeKind::ENUM_TYPE_EXTENSION => function (EnumTypeExtensionNode $def) {
+                    NodeKind::ENUM_TYPE_EXTENSION => function (EnumTypeExtensionNode $def) : string {
                         return $this->join(
                             [
                                 'extend enum',
@@ -424,7 +434,7 @@ class Printer
                         );
                     },
 
-                    NodeKind::INPUT_OBJECT_TYPE_EXTENSION => function (InputObjectTypeExtensionNode $def) {
+                    NodeKind::INPUT_OBJECT_TYPE_EXTENSION => function (InputObjectTypeExtensionNode $def) : string {
                         return $this->join(
                             [
                                 'extend input',
@@ -436,7 +446,7 @@ class Printer
                         );
                     },
 
-                    NodeKind::DIRECTIVE_DEFINITION => $this->addDescription(function (DirectiveDefinitionNode $def) {
+                    NodeKind::DIRECTIVE_DEFINITION => $this->addDescription(function (DirectiveDefinitionNode $def) : string {
                         $noIndent = Utils::every($def->arguments, static function (string $arg) : bool {
                             return strpos($arg, "\n") === false;
                         });
@@ -456,7 +466,7 @@ class Printer
 
     public function addDescription(callable $cb)
     {
-        return function ($node) use ($cb) {
+        return function ($node) use ($cb) : string {
             return $this->join([$node->description, $cb($node)], "\n");
         };
     }
@@ -496,7 +506,7 @@ class Printer
         return $maybeArray ? count($maybeArray) : 0;
     }
 
-    public function join($maybeArray, $separator = '')
+    public function join($maybeArray, $separator = '') : string
     {
         return $maybeArray
             ? implode(