+++ src/Type/Definition/FieldArgument.php
@@ -30,19 +30,14 @@ class FieldArgument
     /** @var mixed[] */
     public $config;
 
-    /** @var InputType&Type */
+    /** @var Type&InputType */
     private $type;
 
-    /**
-     * @param mixed[] $def
-     */
+    /** @param mixed[] $def */
     public function __construct(array $def)
     {
         foreach ($def as $key => $value) {
             switch ($key) {
-                case 'type':
-                    $this->type = $value;
-                    break;
                 case 'name':
                     $this->name = $value;
                     break;
@@ -80,7 +75,17 @@ class FieldArgument
 
     public function getType() : Type
     {
-        return Schema::resolveType($this->type);
+        if (! isset($this->type)) {
+            /**
+             * TODO: replace this phpstan cast with native assert
+             *
+             * @var Type&InputType
+             */
+            $type       = Schema::resolveType($this->config['type']);
+            $this->type = $type;
+        }
+
+        return $this->type;
     }
 
     public function defaultValueExists() : bool
@@ -102,7 +107,7 @@ class FieldArgument
                 sprintf('%s.%s(%s:) %s', $parentType->name, $parentField->name, $this->name, $e->getMessage())
             );
         }
-        $type = $this->type;
+        $type = $this->getType();
         if ($type instanceof WrappingType) {
             $type = $type->getWrappedType(true);
         }